google-oauth/backend-django/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
google-auth==2.14.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
